Refactor Spreadsheet class
