+++ app/spreadsheet_table.rb
@@ -1,0 +1,28 @@
+require './app/table'
+
+class SpreadsheetTable
+  TABLE_LENGTH = 26
+  TABLE_HEIGHT = 100
+
+  def initialize
+    @table = Table.new(TABLE_LENGTH, TABLE_HEIGHT)
+  end
+
+  def get(cell)
+    @table.read_cell(parse_cell_row(cell), parse_cell_column(cell))
+  end
+
+  def put(cell, value)
+    @table.write_cell(parse_cell_row(cell), parse_cell_column(cell), value)
+  end
+
+  private
+
+  def parse_cell_row(cell)
+    Array('A' .. 'Z').find_index(cell[0])
+  end
+
+  def parse_cell_column(cell)
+    cell[1 ..].to_i
+  end
+end
